requirements.txt
re
os
langchain
numpy
pandas
matplotlib
sklearn
torch
torchvision
tqdm
Pillow
opencv-python
requests
pytorch-grad-cam
langchain-chroma
langchain-community
langchain-ollama
openai
gradio